temporarily exclude gpt2 from export
