+++ .github/workflows/test_inf2.yml
@@ -1,4 +1,4 @@
-name: Optimum neuron / Test INF2 INFERENCE
+name: Optimum neuron / Test INF2
 
 on:
   push:
